Fix extraction bug in heap implementation
Whoops. Tests would have caught this, but I don't have those...

This manifested as extraction working not quite right, which meant that
the results had some lower-scoring matches mixed in with the worthy
results.

Noticed while manually testing and comparing quality of the results, and
then used some logging to inspect the results and see what was amiss.