"Modernize" benchmarks
The old Chromium data sets had a lot of `.svn` dot-directories in them,
and given that Command-T generally treats dot-files specially, I wanted
a more representative scenario.

Enter: new Chromium Git repo checkout, but this "only" has 224k files in
it, so something bigger is needed.

I couldn't find any single open source repo big enough, so I combined
the file listings from Chromium, the Linux kernel, and Gecko, bringing
us to a total of about 400k files, which is big enough for now, I guess.

I've also tweaked the iterations count to keep each scenario close to
the 1-second range, freshened the listing in the "command-t" scenario,
and made the necessary changes to the benchmark script for it to
gracefully handle the case where the number of scenarios changes between
runs.