Make in-progress searches faster in LineFinder
Instead of recomputing the "paths" on each key press, keep them around
as long as the finder exists, and keep the finder around for the
duration of the search.

Note that I probably want to be a bit more clever and set up some
autocmds, such that I keep the finder around even longer (until I
switch buffers, or modify the current buffer).