Add primitive first cut at progress reports
This is a bit ugly because scanners don't know about the prompt, only
the controller has a reference to it. We could have passed a reference
to through the matcher to the scanner from the controller, but that
would have been ugly. Ditto if we'd threaded through a callback. So,
instead we opt to just update from the scanner and teach the controller
to mediate by updating the prompt after the scanner has done its thing.

This is ugly, but it's less ugly than the alternatives.

I might want to back off the rate here later on, especially for large
sets.

Doesn't work so well when the find scanner is chewing through a big
result set. I am suspicious of some major buffering going on in
`#readlines` there.