Let "find" scanner really report progress
`#readlines` reads the whole output into memory and then returns an
array, meaning we don't print any progress and then all of a sudden at
the end we print it all.

`#each_line` on the other hand calls our block as we go, which is what
we want.

I'm coding without internet access right now, though, so I am not sure
whether `IO#each_line` is available in all the versions of Ruby that
Command-T currently supports. I will check it out later and may have to
add a fallback.