Add progress animation fanciness
I don't plan on keeping this, but committing it anyway to show the idea.
Basically we don't want to thrash the UI when we're trying to scan, so
we slow down our updates as the result set gets bigger.

The UX isn't great though because it looks like Command-T is slowing
down, even though it isn't. Also, our silly spinner animation goes from
being too fast to being pretty slow.

I'm going to change this to be constant speed based on a time heuristic.