Kill off trailer comments
I've done these for years now as a readability aid, but they are a
liability give my love of copy-pasta coding; they're just another thing
for me to forget to update. Additionally, my classes should be simple
enough that they aren't big enough to warrant this kind of "navigational
aid".