Hack fix for encoding issue
I've found a directory that contains some files in it that cause the
match window to blow up due to non-ASCII characters in their paths.

They are showing up as "ASCII-8BIT" (ie. binary) but only when using the
"watchman" and "ruby" scanners. The "git" and "find" scanners are fine,
and notably the string encoding is "UTF-8" in that case.
`g:CommandTEncoding` doesn't help.

This is a temp fix to make it work, but I am pretty sure there are
better places to put this. I feel bad about mutating the `match` strings
like this. I could perhaps do it in the controller before passing to the
match window. Better still would be to do it when it comes out of the
scanner, although that could incur a speed hit for large hierarchies
(because it would occur on the full set of possible paths, not the
narrowed-down list).

Other "fixes" that I tested that worked:

- Dropping the `u` modifier on the regex.
- Switching back to `chars` instead of `scan(/./mu)` (ie. pre-14fb52d19
  state).

Stuff I tried that didn't work:

- Switching to `\p{Any}` instead of `.` (both with and without the `u`
modifier).