Fork Command-T engine in preparation for v5.0
Going to make Command-T easier to install by running the Ruby bit out of process
and communicating with it by using Vim's new async/job features.

In a normal software project I could make a new branch for this, but Vim users
are used to running from master, and that means that even if I do everything off
on the side, I will eventually have to merge it all back in, so no matter what I
do, breakage will be unacceptable.

So, this means that I have to keep the old code around alongside the new, pretty
much forever. Given that this might happen again in the future I can't go with
any naming scheme that implies "next/current", "old/new", "legacy/modern", or
even "sync/async". Instead, I'm going with codenames with the theme "places in
Middle Earth".

Game-plan then:

1. This phase. Move existing implementation into an engine called "mirkwood".
   Start new "isengard" engine from scratch. Default to "isengard" on
   sufficiently new Vim.
2. There is no step two! (Other than finishing "isengard".)