Show that it is possible to require the C extension
The failure mode here is not great. If I `raise` an error, Vim effectively spins
forever, printing endlessly to the `clog`:

```
0.000765 on 0: Blocking NL read, timeout: 2000 msec
0.000768 on 0: Waiting for up to 2000 msec
0.044705 ERR on 0: channel_read_block(): Cannot read from channel, will close it soon
0.044716 on 0: Waiting for up to 2000 msec
0.044722 ERR on 0: channel_read_block(): Cannot read from channel, will close it soon
0.044725 on 0: Waiting for up to 2000 msec
```

Not sure why that is, nor where the retry behavior is coming from, but I have to
be very careful of printing anything to STDOUT/STDERR I think other than the
messages I intend to send to Vim explicitly.