Move Command-T files deeper
We want to have a more conventional structure for the gem, as I will ship the
daemon process with it. That means we need a `bin` directory, so dividing things
more conventionally among `lib`, `bin` and `ext`.

This is completely untested and very likely broken, but I'll test and fix as I
go along.