Speed up file scanners.
This makes a couple of changes to speed up file scanners. The biggest
component is replacing some vimscript to check if a file is excluded to
using Ruby's binding to fnmatch. This saves a lot of work and speeds up
scanning significantly.

Other changes were mostly removing memory allocations and other small
changes.