Add option for custom scanning command.
This option provides the user with ultimate flexibility. They have the
choice to provide any shell command, script or pipeline in order to
generate a file list.

To give the command more context in the future variables can be set before
calling the user command in the shell to pass more information to the external
command. This allows future backwards compatible extension.

The only drawback I see is that the terminator is "static" for example if
a user has a script that switches the technique used based on context
(for example current path) it will always have to use the same
terminator. Luckily the workaround is simple by using `| tr '\n' '\0'`
to convert newlines (or whatever the alternative separator is) to nulls
and setting CommandT to always use nulls. This has a slight CPU and
memory overhead but since it runs in parallel it should be insignificant
on a multi-core machine.