Fix turning */foo/* wildignore into a regexp
I was trying to get rid of files inside dist folders in my command-t
search results. So I added this to my wildignore setting for vim:

  */dist/*

...which didn't seem to have any effect. I'm using watchman, so at first
I thought that there might some cache that needs to be invalidated. But
`watchman watch-del <path>` didn't help.

I decided to change the wildignore pattern to `*/dist`. This worked, and
didn't require any cache invalidation.

Digging through the code, I found a simple way to reproduce the but in
an rspec test. So I decided to fix it, by
- Flipping the order between the `*/something` pattern and the
`*/something/*` pattern, to deal with the most complicated pattern
first.
- Changing the matching pattern to look for a slash instead of a star.
- Changing the resulting regexp to include the slash plus anything
after.