Fix problem opening identically named buffers with `:sbuffer`
Closes: https://github.com/wincent/command-t/issues/295

Just say we open "win32/capturer.rs" and then try to open "capturer.rs".

For the first file, we'll fall through to the `else` branch in
`commandt#GotoOrOpen` and open it with `:edit`. For the second file, our
`buwinnr()` check will report that the buffer is already open, so we'll
try to open it with `:sbuffer` which causes us to go back to the
previous file instead of opening the new one that we wanted.

The fix here is to do a stricter match in `bufwinnr()`. In this way,
when we try to open the second file ("capturer.rs") our
check(`bufwinnr("^capturer.rs$")`) will fail, and we'll open the new
file correctly with `:edit`.

I've tested this manually with a bunch of permutations of splits, tabs,
and file orderings, and I don't believe this breaks any of the existing
functionality.