Make Ctrl-W match zsh-ish behavior
As pointed out in https://github.com/wincent/command-t/issues/301 Ctrl-W
would not delete past punctuation characters.

I did some testing in zsh to see what that does and updated the regex to
match:

   foo bar.baz        (deletes "baz" then "bar.")
   foo bar....baz     (deletes "baz" then "bar....")
      indented        (deletes "indented" but won't delete leading ws)

Note that unlike the old pattern, it is possible for this one to not
match, so I added an `if` guard.

Should probably have tests for this.

Closes: https://github.com/wincent/command-t/issues/301