Fix regression trying to switch to previously opened file in other tab
This is a regression introduced in 85949ce7.

That one attempted to fix #298 by applying a similar fix to the one used
in #295. However, it only worked by luck: `bufexists()` doesn't support
patterns or use of "$"/"^" to anchor the match like `bufwinnr()` does.
This meant that adding those anchors made it always false.

That fixed the unwanted splits in #298, but it meant that you couldn't
go back to an existing file in another tab.

The fix is pretty straightforward: switch from `bufexists()` to
`bufname()`, because the latter does support anchoring. As far as I can
tell, this fixes #306 without breaking the scenario detailed in #298.
The one gotcha here is that we can't rely on the truthiness of the
non-empty string returned by `bufname()` here, because crazy old
Vimscript coerces also strings to `0`; we therefore have to use an
explicit check against `''`.

Closes: https://github.com/wincent/command-t/issues/306