Add new file scanner: rg (ripgrep)
It works very similarly to the find(1) scanner but uses ripgrep[1] as
the underlying tool.  Ripgrep is primarily a replacement for grep(1) but
can be used in a way similar to find(1).  It respects gitignore rules,
which can be an advantage over the find(1) scanner, but it only lists
files under $PWD, which can be an advantage over the git(1) scanner.

[1]: https://github.com/BurntSushi/ripgrep