Work around pathologically slow buffer scanning in Neovim
In Neovim, unlisted buffers are returned to the Ruby layer. If you run
Ferret's :Ack, for example, you might find yourself with a large number
of those (as `:h unlisted-buffer` says, Vim uses these to track things
like file names). As such, it is too easy to wind up with a *huge*
number of unlisted buffers, and the scanner takes forever.

This is exacerbated by the scanner being called on every keypress.

Fix the slow processing by avoiding use of `VIM::Buffer` entirely, and
instead call out to native Vim commands.

This alone makes it usable again. In a follow-up commit I'll do the
"right" thing and make it not call the scanner repeatedly during a
single invocation.