Fix CommandTMRU degenerate performance under Neovim
Same as we did with CommandTBuffer, but a deeper refactor. Switches the
MRU module to track buffer numbers instead of full-blown buffer objects.
If nothing else, this means we'll not be as likely to accumulate garbage
over time.

The perf is fine now, despite the O(n^2) logic in the unused buffer loop
(the `#used?` method scans the entire stack, and we call it once for
every buffer in the list) because n is usually small (even 200 * 200
comparisons here is plenty fast). I'll still replace that with something
O(n) though as a follow-up, because otherwise it offends my
sensibilities a bit.