docs: add tips for compiling when using rvm and brew
Homebrew now has ruby listed as a dependency for both vim and
macvim, which means there are version conflicts together with
RVM (and likely rbenv). I have confirmed that the comment at
https://github.com/wincent/command-t/issues/341#issuecomment-522216390
fixes this issue (for my computer running Catalina and macvim from
homebrew), and am adding these to the instructions so they are easier
for others to find.