refactor: tweak per `:h wildcard` definition
Given that `:h wildcard` says this about "*" and friends:

        ?       matches one character
        *       matches anything, including nothing
        **      matches anything, including nothing, recurses into directories
        [abc]   match 'a', 'b' or 'c'

it means that a pattern like "*~" is supposed to match not only
"file~" but also "~" (and yes, this is a valid filename even though
you'd be crazy to make it).

So, tweaked the regex to match that by dropping the requirement for
any preceding match at all. I simplified the conditional too; now,
instead of treating the first position different from the following
positions, it just treats them all alike. In other words, the previous
version allowed a pattern like "*a.b"; now, it just allows "*a", "*ab"
etc.