chore: battle bit-rot in gem dependency graph
Because GitHub is nagging me:

    Upgrade nokogiri to version 1.10.8 or later. For example:

    gem "nokogiri", ">= 1.10.8"

    CVE-2020-7595

    moderate severity

    Vulnerable versions: < 1.10.8

    Patched version: 1.10.8

    xmlStringLenDecodeEntities in parser.c in libxml2 2.9.10 has an
    infinite loop in a certain end-of-file situation.  The Nokogiri
    RubyGem has patched it's vendored copy of libxml2 in order to
    prevent this issue from affecting nokogiri.

And this dependency brings me net negative value. I can just prepare
and upload releases manually, without mechanize. It's not like I do it
often anyway; anybody using pretty much any Vim plug-in nowadays just
points their plug-in manager at the `master` branch of a Git repo --
"releases" are almost meaningless in this context, and I'm only really
going to do it from time to time to keep the plug-in page on
www.vim.org alive.

Note also that I'm not going to both uploading to S3 anymore. I have
nowhere linking to it.