refactor: show callability of `commandt_calculate_match()`
So this one is a bit of a mess. You can see I started by removing
Ruby-specific bits of matcher.c, but I started to get a bit worried that
I was going to get to the point that everything compiled but the code
didn't actually work at all, so I scaled back to a simpler test:
basically just showing I can call `commandt_calculate_match()` and get a
result back (I can).

As noted inline, it's just a demo. The function is for internal use only
anyway, and it takes a pointer to an "out" param, which I don't know how
to model in Lua's FFI. Will see what I can find though, as it might be a
useful utility to have.