fix: don't error if watchman dies with SIGABRT
On macOS, my `watchman` installation was busted such that any attempt to
launch it from the command-line was dying with this:

    dyld[13871]: Library not loaded: /usr/local/opt/icu4c/lib/libicudata.69.dylib
  Referenced from: /usr/local/Cellar/boost/1.78.0/lib/libboost_regex-mt.dylib
  Reason: tried: '/usr/local/opt/icu4c/lib/libicudata.69.dylib' (no such file), '/usr/local/lib/libicudata.69.dylib' (no such file), '/usr/lib/libicudata.69.dylib' (no such file), '/usr/local/Cellar/icu4c/70.1/lib/libicudata.69.dylib' (no such file), '/usr/local/lib/libicudata.69.dylib' (no such file), '/usr/lib/libicudata.69.dylib' (no such file)
    zsh: abort      watchman

Inside Command-T, this manifests as follows:

- `$?` is `<Process::Status: pid 13852 SIGABRT (signal 6)>`
- `$?.exitstatus` is `nil`
- `$?.exitstatus.zero?` explodes with `NoMethodError (undefined method
  `zero?' for nil:NilClass)`

Mitigation here is to use safe navigation operator (`&.`), which was
added in Ruby 2.3.0 (released in December 2015, so I am comfortable
using this, nearly 7 years later... if I get any bug reports I can
always rewrite this in a backwards compatible way).

For reference, fixing my `watchman` install was a question of doing a
`brew install watchman` again.