feat: use watchman's `watch-project` if available
If watchman is bersion 3.3 or later, use `watch-project` rather than
the deprecated `watch`. The main advantage is eliminating redundant
nested watches (i.e.  where one is a subtree of another), which reduces
RAM and CPU usage in the `watchman` daemon and eliminates the indexing
delay when invoking Command-T after moving down the directory hierarchy.

Although `watch-project` was introduced in 3.1 it is awkward/inefficient
to use without the `relative_root` query which was introduced in 3.3.
3.3 was released in June 2015, so most users should have a newer version
but backward compatibility with older versions is maintained.

Fixes #389

Tested:
* Tested the new code under various scenarios by invoking Command-T in my homedir
  and various subdirectories, as well as other paths (/tmp).
* Manually tested the fallback path to `watch`, which is unchanged from existing code.
* Tested various version numbers by manually simulating the output of `--version`
    * If the version number is missing or malformed, it will end up as version 0.0
    * test cases: nil, "", "foo.bar", "-5.2", "1", "3.2", 3.3.449", "4", "2022.06.13.00"