refactor: tweak demo function to show it taking an array of str_t
It returns a count, and along the way prints details of the structs that
were passed into it, just to show that they actually made it through ok.

At some point will have to benchmark those:

    ffi.new('str_t', {'stuff', 5, 5})

calls against just passing in strings and calling `strlen()` on them. I
added a TODO for that.