fix(lua): avoid EXC_BAD_ACCESS due to uninitialized struct fields
Because we didn't fully initialize the `matcher_t` struct, it could have
garbage in the `haystacks` field when we create it, which in turn would
cause our lazy initialization not to run, and the relevant pointers to
be invalid. And we'd then blow up as soon as we tried to access them, as
seen in the parent commit message.