refactor(lua): rationalize types
Use `size_t` for things that are byte counts, or for passing to APIs
that expect `size_t` params.

For things that are object counts, prefer `unsigned`.

There is one part of this refactoring that might be a bit buggy: in
`commandt_score()` (`score.c`), I added a short circuit where we break
out of the `while` loop if it is impossible for the rest of the needle
to be found in the rest of the haystack:

    while (haystack_idx >= needle_idx) {

The problem with this is that, right after the `while`, we store the
computed mask:

    if (compute_bitmasks) {
        haystack->bitmask = mask;
    }

Evidently, if we short-circuit the loop, then the mask won't actually be
complete. I have a vague recollection that I might have tried a
refactoring like this in the past, and ended up discarding it for
exactly the same reason. Whatever; in the next commit I will plug this
hole.