perf(lua): invert order of wrapping in comparator functions
As noted in the parent commit, seems like it is a good idea to optimize
for the common case (doing zillions of heap comparisons during matching)
as opposed to the small-N case (doing a sort over a limited set of
results at the end).

So, instead of having every heap operation pay the overhead of an extra
function call (through the wrapper, to the qsort-oriented comparator),
do it the other way around: the heap gets the simple comparator, and the
sorting gets the wrapper.