fix(lua): make it possible to stop nvim-cmp from showing a popup
It gets in the way in the prompt window.

With the filetype in place, you can create a
`ftplugin/CommandTPrompt.lua` file with contents something like this:

    -- Stop nvim-cmp from showing autocomplete popup while typing into
    -- CommandTPrompt window.

    local has_cmp, cmp = pcall(require, 'cmp')
    if not has_cmp then
      return
    end

    if not wincent.g.CommandTPrompt then
      wincent.g.CommandTPrompt = {registered = true}
      cmp.setup.filetype('CommandTPrompt', {
        sources = cmp.config.sources({}),
      })
    end

as seen here:

    https://github.com/wincent/wincent/commit/3fed25eb7de753b5fa9b61a7dffcdf0c47258239

In short, it tells nvim-cmp to use an empty sources list in this
filetype. The `wincent.g.CommandTPrompt` thing is just a guard to make
sure I do this setup only once.

Resorted to this after my other approaches failed to yield a result:

    vim.api.nvim_buf_set_option(buffer, 'complete', '')
    vim.api.nvim_set_option('completeopt', '')