fix(lua): fix off-by-one in key detection loop
As hinted at in parent commit message, the problem is that:

    sizeof("string")

includes the trailing NULL byte, but:

    strlen("string")

does not. This meant that we were failing to find the expected "watch"
key.

With the fix for the off-by-one in place, our Lua-side logging:

    print(vim.inspect(result))
    print(vim.inspect(result['watch']))
    print(vim.inspect(result['relative_path']))

prints:

    cdata<struct 141 *>: 0x600000050150
    cdata<const char *>: 0x600002602460
    cdata<const char *>: NULL

So, I'm including a small tweak to make that instead print:

    {
      watch = "/Users/wincent/code/wincent/aspects/nvim/files/.config/nvim/pack/bundle/opt/command-t"
    }