perf(lua): force C library to load before taking timings
This explains the errant timings mentioned in the parent commit:

Basically, we were doing this:

1. Record CPU time.
2. Record wall-clock time, incurring hit of loading C library for first
   time.
3. Do work.
4. Record CPU time again.
5. Record wall-clock time again.

So, the delta between "1" and "4" was artificially inflated because
it included the cost of loading the C library. Ironically, the delta
between "2" and "5" was _not_ inflated, because the timing info is
actually obtained and returned from the first `commandt.epoch()` call
only after the C library has done loading.

With the change in this commit, the times line up nicely:

    elapsed (CPU): 0.000157
    elapsed (wall): 0.000157