feat(lua): add historical comparison to benchmarking
Lots of code here, so it might have bugs (I'll check it tomorrow), but
this prints out stuff like this:

    Summary of cpu time and (wall time):

                        best    avg      sd     +/-      p    (best)  (avg)    (sd)     +/-      p
    chromium (subset) 2.81499 4.07109 1.72634 [+4.1%] 0.0005 0.52112 0.85807 0.61572 [+14.8%] 0.0005
           big (400k) 5.38775 6.75959 1.92037 [+1.9%] 0.0005 0.74014 1.08010 0.64347 [+13.4%] 0.0005
     chromium (whole) 3.31954 4.36914 1.42175 [+1.5%]  0.005 0.45263 0.72852 0.51081 [+18.8%] 0.0005
         pathological 0.45680 0.70984 0.35737 [+5.3%] 0.0005 0.45725 0.71781 0.37912  [+6.3%] 0.0005
            command-t 0.41544 0.64982 0.33837 [+5.1%] 0.0005 0.41565 0.65967 0.37454  [+6.3%] 0.0005

based on data logged to `data/wincent/benchmark/log.lua`.

It's hot here, so the above is the result of running twice in a row with
`TIMES=10`, and we see a real (as indicated by the p values) slow down
due to thermal throttling. If I let the machine cool down for a minute
and then run again, I see this:

    Summary of cpu time and (wall time):

                        best    avg      sd      +/-      p    (best)  (avg)    (sd)     +/-      p
    chromium (subset) 2.82297 3.00507 0.31863 [-35.5%] 0.0005 0.54307 0.58483 0.07685 [-46.7%] 0.0005
           big (400k) 4.98921 5.39459 0.63969 [-25.3%] 0.0005 0.76641 0.85499 0.30155 [-26.3%] 0.0005
     chromium (whole) 3.25751 3.55554 0.45382 [-22.9%] 0.0005 0.47375 0.53356 0.10181 [-36.5%] 0.0005
         pathological 0.45641 0.47504 0.05603 [-49.4%] 0.0005 0.45725 0.48064 0.09039 [-49.3%] 0.0005
            command-t 0.42371 0.43316 0.02380 [-50.0%] 0.0005 0.42432 0.43432 0.02498 [-51.9%] 0.0005

Of course, with the hot weather, it's going to be really hard to get
sensible comparisons for a while (it's 40 degrees celcius, 100+ degrees
farenheit). In order to stay cool we have to run only a small number of
times, but in order to get high signal data, we have to run lots of
times... so...

Still to do, as noted in the comment, sort the table so it shows the
rows in the same order as we use elsewhere.