perf(lua): show progress counts during benchmarking
Now we show:

    Rehearsal 9 of 20          total         wall
    ---------------------------------------------
    pathological            0.461023   (0.461713)
    command-t               0.440428   (0.440672)
    chromium (subset)       3.089496   (0.569606)
    chromium (whole)        3.746735   (0.511079)
    big (400k)              5.659923   (0.761181)

    Final 9 of 20              total         wall
    ---------------------------------------------
    pathological            0.484537   (0.484941)
    command-t               0.413163   (0.413298)
    chromium (subset)       3.062933   (0.587998)
    chromium (whole)        3.715158   (0.513318)
    big (400k)              5.639722   (0.804543)

which is a bit easier to read, and gives an indication of how far to go
in a long run.

Note the header layout is a bit fragile because it doesn't take into
consideration the cell widths. Further down the file we do a '%-22s'
format marker that assumes our labels are never wider than that.

Probably won't break, as I am not planning on changing labels or adding
more, but still wanted to at least note it here.