fix(lua): fix build spew on Debian
On a Debian box, I see Make using `/bin/sh`, which is actually
a symlink to `/bin/dash`, and it hates the `[[` syntax:

    /bin/sh: 1: [[: not found

This is harmless because the build proceeds anyway, but we should avoid
the spew.

Force it to use Bash. (I could also pick some more POSIX-y syntax, if I
discover a reason to not use Bash.)

This wasn't a problem on Arch or Darwin, where `/bin/sh` groks `[[`
just fine (on this Darwin machine where I am now, I'm not actually sure
what `sh` is though; `sh --help` shows Bash version info, but it's not a
symlink and the file size is different from `/bin/bash` 🤷).