refactor(lua): move most files into "private" subdirectories
To make it clear what the boundary between "public" and
"internal" APIs is. The idea is that anything you can
see with `require'wincent.commandt'` is fair game, but
`require'wincent.commandt.private.$anything'` is not.

This is a pattern I'm already using in Command-T (see
`autoload/commandt/private.vim`), and in other plug-ins; eg:

    https://github.com/wincent/ferret/tree/3d064304876941e4197db6b4264db6b72bd9f83d/autoload/ferret

Additionally, we're assuming that the _other_ directories where we have
Lua, are off limits:

- `bin/`: This is for development purposes (testing and benchmarking
  scripts); and:
- `data/`: This contains benchmarking inputs and logs. If anybody does a
  `require'wincent.benchmark'` or a `require'wincent.benchmark.log'`,
  then that's on them; it doesn't seem worth moving _those_ under a
  private folder because the _entire_ `bin/` and `data/` directories are
  supposedly private.
- Likewise for test files (currently only two of those:
  `lua/wincent/commandt/private/scanner/test/init.lua` and
  `lua/wincent/commandt/test/matcher.lua`) aren't intended to be
  `require`-d by anybody but the test script, and the name should make
  that obvious so I'm leaving them as-is.

The C code now all lives in:

  lua/wincent/commandt/lib

Again, that's effectively private, but I don't think we need to make
that obvious in the file path because you can't just go and require the
library willy-nilly (you need the FFI set-up, which you can only get
with `require'wincent.commandt.private.lib'`.

Still to do:

- Make sure there is nothing in `wincent.commandt` that should actually
  be hidden behind a layer of indirection. At _least_ `epoch()` should
  be move out of it, but I will check for other things too.