perf(lua): switch to "slab" allocation mode in Watchman scanner
Because I'll be wanting to use this in some other upcoming scanners too.

It's messy, for sure, but the idea is that, instead of having a scanner
manage a bunch of pointers to `str_r` structs that themselves have
pointers to real string storage (`const char *`), we store the `str_r`
structs directly in an array. In this way we can allocate them all and
once, and free them all at once.

Additionally, when we already have a slab of memory allocated for
something like a Watchman response, and that slab contains the string
contents, we can just keep that slab around (the overhead is low as
there is not _that_ much metadata in the response relative to the string
contents, and even if there is overhead, being able to allocate just one
slab for the response is _huge_, compare with hundreds of thousands of
tiny allocations).

It's hard to de reliable benchmarking with such hot weather, but the
numbers at least don't seem to get worse as a result of this change, and
maybe they even get quite a bit better...

As usual, these are for 20 runs, "cpu" and "(wall)":

                        best    avg      sd      +/-     p    (best)  (avg)    (sd)     +/-     p
         pathological 0.28742 0.30130 0.10830 [-14.8%]  0.05 0.28741 0.30216 0.11403 [-14.5%]  0.05
            command-t 0.24418 0.25537 0.09166  [+4.7%] 0.005 0.24417 0.25563 0.09535  [+4.6%] 0.005
    chromium (subset) 1.62233 1.64373 0.12668  [-7.9%] 0.025 0.41818 0.43359 0.08519  [-4.4%] 0.025
     chromium (whole) 1.88715 1.92114 0.11793 [-16.2%] 0.025 0.30404 0.32124 0.08993 [-17.6%] 0.005
           big (400k) 2.92630 2.97203 0.17515 [-21.6%]  0.05 0.46431 0.48693 0.09316 [-19.3%]  0.01
                total 7.00522 7.09357 0.31774 [-15.7%]  0.05 1.72976 1.79955 0.18531 [-11.2%]  0.05