perf(lua): add initial version of scanner benchmarks
Obviously these are very rough, and I have some kind of resource
exhaustion or coordination bug, because the run aborts before getting to
the end. Also, I'm not sure why the Watchman scanner seems so slow in
this environment.

Excerpt:

    Rehearsal 2 of 20             cpu         wall
    ----------------------------------------------
    git                       0.02951    (0.67236)
    rg                        0.02989    (0.82734)
    watchman                  0.00049    (0.73985)
    total                     0.05989    (2.23955)

    Final 2 of 20                 cpu         wall
    ----------------------------------------------
    git                       0.02889    (0.67417)
    rg                        0.03038    (0.83334)
    watchman                  0.00041    (0.74906)
    total                     0.05968    (2.25658)

    Rehearsal 3 of 20             cpu         wall
    ----------------------------------------------
    git                       0.03111    (0.68687)
    zsh: abort      bin/benchmarks/scanner.lua

ie. seems to be dying somewhere around the `rg` variant of the 3rd
rehearsal. Also note how the `watchman` variant shows super low CPU
time, but relatively high wall-clock time, despite the fact that I've
got it dialled down to run 5 times (instead of 100, like for the other
scanners).

I think the most likely cause is that GC isn't kicking in quickly enough
for us to return allocated resources to the pool (or maybe I have a
leak), and so somewhere in there there'll be a `malloc()` or an `mmap()`
that fails, or perhaps I'm leaking file descriptors and exhausting them
etc. Will look into it later.