docs: update HISTORY section with notes from 5.0.5 release
This is a cherry pick of 806b20304db828e159fb886087a3b8dfbd922ec1
("docs: update HISTORY section prior to release", Jul 25 2022), on the
"5-x-release" branch. That may well be the last release in the 5.x
series for a long, long time. "pu" will go to "next"/"main" soon, and
we'll call it 6.0.0-a.0 (or maybe 6.0.0-a.1; I'm not sure yet).