feat(lua): add more movement bindings
In normal mode, `gg`, `G`, `H`, `M`, `L` all now do what you would
expect. This wasn't something the Ruby implementation could do, because
there was no concept of "normal" mode (the prompt wasn't a buffer but a
synthetic thing that intercepted key presses via mappings and then
hackily echoed into the command line area).

As this would cause the number of named callbacks to explode, I reframed
the existing ones like `on_next`, `on_previous` etc in terms of a single
callback called `on_select` which takes params like this:

- `{ absolute = 1 }`: select first item.
- `{ absolute = -1 }`: select last item.
- `{ absolute = 0 }`: select middle item.
- `{ absolute = n }`: select numbered item
  (not actually used other than as noted above).
- `{ absolute = -n }`: select item counting back from end
  (not actually used other than as noted above).
- `{ relative = 1 }`: select next item.
- `{ relative = -1 }`: select previous item.
- `{ relative = n }`: select nth next item
  (not actually used other than as noted above).
- `{ relative = -n }`: select nth previous item
  (not actually used other than as noted above).

In order to make this "select" terminology available, I renamed the
exist "select" (which in Ruby was "accept selection") to "open".

Not yet implemented but added to the list:

- `<C-u>` to clear to the left in normal mode (already works in insert
  mode).