refactor(lua): rename `case_sensitive` to `ignore_case`
And invert the sense.

Next step will be to add a `smart_case` option too.

We need to do this because of the way the matcher is set up and run. We
pass in the settings at construction time, and we can't know what to do
with the needle at that time (ie. because of smartcase), only at
runtime. In other words, we have to figure that out in
`commandt_matcher_run()`.