feat(lua): wire up the native file scanner
In a very pleasing refactor, I can get rid of my own directory traversal
code and use `fts_open()` and friends instead; this is much cleaner,
less code, and should nicely handle edge cases for me. In testing, it
seems fast enough (I'm just messing with the Linux kernel repo, which is
the largest one I have handy on this machine), and some comments on the
web suggest that it is a pretty good baseline in terms of speed (note:
these comments are macOS specific):

    https://mobile.twitter.com/tempelorg/status/1120287110429335553

For some rough perf numbers, here is a snapshot, which I have edited by
hand to add a "runs" count column to make it easier to gauge the
relative speed:

    Final                         cpu         wall     runs
    -------------------------------------------------------
    buffer                    0.02473    (0.02476)    10000
    file                      0.06013    (0.06054)      100
    find                      0.01891    (0.20348)      100
    git                       0.01871    (0.22548)      100
    rg                        0.01740    (0.56418)      100
    watchman                  0.00024    (0.00131)        5
    total                     0.14012    (1.07975)    10405

Here we see that the built-in `file` scan is about 3x fast than both
`find` and `git`, which we can largely attribute to forking overhead.
On a small repo like this one, `rg` is slower (3x slower again than
`find` and `git`); it must have some fixed cost set-up overhead, because
I would expect it to beat `find` (maybe not `git` though)... maybe on
larger repos it would... As for `watchman`, as you can see I have that
one turned way down to only run 5 iterations because it had seemed slow
in prior testing; at least now though, it doesn't look to be slow at
all; if we multiply its time by 20 to get the equivalent of 100 runs, we
get an estimated time of (0.02620), which would be about 3x faster than
`file`.