chore(lua): tweak the benchmark run count
As I mentioned in parent commit, the "watchman" suite was artificially
shortened because it had seemed slow. For whatever reason, it doesn't
seem slow now, so I am bumping it back up.

    Summary of cpu time and (wall time):

               best    avg      sd     +/-   p   (best)    (avg)      (sd)     +/-   p
      buffer 0.02457 0.02457 0.00000 [-1.4%]   (0.02461) (0.02461) (0.00000) [-1.3%]
        file 0.05938 0.05938 0.00000 [-1.3%]   (0.05981) (0.05981) (0.00000) [-0.9%]
        find 0.01901 0.01901 0.00000 [-0.5%]   (0.21205) (0.21205) (0.00000) [+0.8%]
         git 0.01897 0.01897 0.00000 [+0.1%]   (0.22607) (0.22607) (0.00000) [+0.1%]
          rg 0.01721 0.01721 0.00000 [+2.4%]   (0.56951) (0.56951) (0.00000) [-0.1%]
    watchman 0.00188 0.00188 0.00000 [-1.1%]   (0.01833) (0.01833) (0.00000) [+0.9%]
       total 0.14101 0.14101 0.00000 [-0.5%]   (1.11038) (1.11038) (0.00000) [+0.1%]

That is, `file` is around 4x faster than `find` or `git`, which are
themselves about 2.5x faster than `rg`, but `watchman` beats them all:
it is about 3x faster than `file`. I'll have to see if these results
replicate on another machine.

I still have the problem though that I can't run the whole benchmark
with a high `TIMES` count, so there is definitely a lurking bug in there
somewhere (perhaps exhausting file descriptors, or not leaking memory,
or GC is not kicking in fast enough to accommodate so many allocations
in a small interval).