refactor(lua): rationalize our external symbols
With the help of `nm -gD commandt.so`, we can see all external symbols
in the library.

We want everything to be prefixed with `commandt_` to avoid
collisions.  But I don't want to write that everywhere, so... add
some macros so I can keep writing things like `heap_new()` instead
of `commandt_heap_new()`. I am happy for some things to really have
`commandt_` in the name (ie. in the source, and have to write it out
every time), because they are totally specific to Command-T (things
like scanners, matchers etc), but things that are generic, like heaps,
strings, and functions like `xmalloc()` and so on, I don't want those to
appear with the prefix in the source, because, conceptually, they are
things that I could at least in principle use in another project. So,
external symbols have prefixes, source code generally does not, except
when it does.

Do I love that the macro is obscuring the reality of what the actual
symbol names in the built library are going to be when you're looking
at the source? Not really, but I prefer it to the alternative, which is
_not_ obscuring that reality. Them's the breaks with C.

Conversely, in the (Lua) `lib` layer we don't need the prefixes so I
have dropped them from everywhere. As in, why would you want a redundant
prefix in something like this?

    require('wincent.commandt.private.lib').epoch()