feat(lua): pass options through to Git scanner
Note that unlike in the Ruby version, we are relying on
`--recurse-submodules` to do the heavy-lifting for us. This setting is
incompatible with `--others`. Ruby got around this by doing its own
separate invocations of `git-ls-files` in the submodules, but that's not
an option for us here because of the way scanners work with a single
command invocation (we could write a script/pipeline, I guess, and
execute that, but I don't think I want the complexity).

The idea, instead, is that if you want to default to recursing into
submodules and occasionally want to open an untracked file, you switch
to using a different finder (unlike the Ruby version, that's easier,
because each finder has a separate command; before, you actually had to
change a setting).

Obviously, the options validation is super verbose, but I want to write
it all out a few times to see the emergent pattern before I go and
prematurely abstract it. Maybe in an imminent commit.