test: avoid another deprecation warning
    spec/command-t/watchman/utils_spec.rb:17: warning: constant
    ::Fixnum is deprecated

According to:

    https://bugs.ruby-lang.org/issues/12005

> CRuby has two Integer classes, Fixnum and Bignum.
> But it is implementation detail.
> They should be seen as a single class Integer

Unification happened in Ruby 2.4.0, which is approximately $zillion
years ago (well, 25 Dec 2016):

    https://www.ruby-lang.org/en/news/2016/12/25/ruby-2-4-0-released/

I doubt anybody running these tests is on such an old version at this
point, so it seems safe to switch to `Integer`.

Removed the comments because I am not certain whether we're going to hit
that first `if` branch's body any more, and I'd rather have no comments
than something inaccurate or misleading.