refactor: mostly clean out Rakefile
We're not going to be cutting Ruby releases from this branch any more
(see `5-x-release` for that), so most of this file becomes irrelevant.
As noted in the "TODO" comments, I want to remove the release-y bits out
of the Rakefile in a future commit (at which point all it will really be
for is to run the specs, and to be honest, if that just amounts to
`bundle exec spec` I may end up getting rid of the whole file).