docs: remove reference to `g:CommandTSuppressRubyDeprecationWarning`
I never implemented this because I realized that if you are going to
set:

    let g:CommandTSuppressRubyDeprecationWarning=1

you may as well just do:

    let g:CommandTPreferredImplementation='ruby'

so, the suppression setting is pointless. I also abandoned my plans of
showing the message only once (ie. by writing something to disk) as not
worth the complexity — too many edge cases.