fix: use smaller memory allocation values on 32-bit systems
To deal with the aborts reported in:

    https://github.com/wincent/command-t/issues/399

My theory was that on a 32-bit machine an allocation was failing due to
the rather generous sizes involved, given that the DEBUG build showed
the tests passing, and some allocations were recorded in the
`commandt-debug.log`

We tried with smaller sizes (1/4 of the previous sizes) and the aborts
went away. These sizes are still ridiculously large. The smaller sizes
correspond up to 32 million candidate strings (the largest projects I've
ever worked with or even heard of had anywhere between 1.5 M and 4 M
strings; I expect that anybody working on anything larger than that is
very, very likely running on a 64-bit architecture), and a string
storage slab of 32 G (the way the math works out here is if you had 32 M
strings of length 1K you'd fill up all the space).

So, in this commit, we do some basic autodetection based on `getconf
LONG_BIT` (which is POSIX, so should be relatively portable, even if it
isn't super informative) and use that to decide how big the numbers
should be.

We may still want to make this overridable by users, or otherwise more
sophisticated, but until I see a demand for that, I'll hold off.

The `compile_flags.txt` file is so that `clangd` can know what the
preprocessor options are going to resemble (obviously, these are
hard-coded, so really are just intended to keep the LSP from complaining
about undeclared identifiers).