style: use uniform indents in Makefile
As it says here:

    https://stackoverflow.com/a/28720186/2103996

the only place tabs matter is in the rules, but we may as well use tabs
everywhere. We could also use spaces; it would work just the same.
Indenting does make it more readable, although it does bust the syntax
highlighting in Vim a little bit, so I guess _somebody_ out there thinks
you should not be using indents for non-rule parts of the Makefile.

I would use something like git.git's Makefile as a guide, but it uses
both indented and unindented conditional bodies. 🤷