fix: avoid `invalid option '%-100' to 'format'` and similar
As described in the related issue, I didn't have a reliable repro, but I
have seen this error message on occasion:

> Error executing lua callback:
> ...command-t/lua/wincent/commandt/private/match_listing.lua:65:
> invalid option '%-100' to 'format'

I first suspected a bad cached value, and that may be the case, but
either way, the arithmetic here seems off, as it will allow a bad format
to occur whenever `width` is, say, 103 (`width` usually comes from
a `Window` instance, with `vim.o.columns` used as a fallback).

I was finally able to repro by setting `margin` to `1` and then
adjusting my window size until `'columns'` was (I think) 107.

Closes: https://github.com/wincent/command-t/issues/415