refactor: merge lists by overwriting instead of fusing
We'll need this shortly to allow users to override a default setting
that is a list; eg.

- given default value: `{ 'a', 'b', 'c' }`
- and override: `{ 'x', 'y' }`
- we want final result: `{ 'x', 'y' }`
- and not: `{ 'x', 'y', 'c' }`

So, the behavior is to merge table-like tables, but overwrite list-like
lists. This means you can:

- Easily add to a table-like table.
- But also remove from a table-like table (by setting a value to `false`).
- Easily replace a list-like table.
- But also remove from or add to a list-like table by obtaining the
  default value and combining that however you wish with you new values.

To illustrate that last point, say you want to _add_ to a default list
and/or maybe remove from it without spelling the whole thing out, you
can:

1.  Grab the original value:

      local items = require('wincent.commandt').default_options().some_list

2.  Append your new item:

      table.insert(items, "extra")

3.  Remove the first item:

      table.remove(items, 1)

4.  Do something with the list...

      require('wincent.commandt').setup({
        some_list = items,
      })

This works because, for now at least, we're not using complex lists in
our config (eg. a list of tables where we conceivably might want to
merge the tables inside the list). In theory that kind of scenario can
be handled along the same lines as illustrated above, but it could get
quite cumbersome. If we need to handle that kind of config in the
future, we can re-evaluate this decision.