refactor: lay groundwork for "traverse" and "root_markers" settings
With this we can find the root along the same lines as we did in the
Ruby implementation with:

- g:CommandTTraverseSCM
- g:CommandTSCMDirectories

I'm not actually wiring it up yet because our scanners have some
assumptions baked into them that are incompatible. ie. They don't cope
well when passed a big-long absolute path, or even an upwards-pointing
relative path, although downwards-pointing relative paths work ok. That
is, a subdir path like "foo/bar" works pretty well, but "../.." is a
mess.