fix: avoid falling back to fallback unnecessarily
As noted here:

- https://github.com/wincent/command-t/issues/420#issuecomment-1763201612

It's possible to hit the max file limit (coming soon to a repo near you)
but have all the results hidden because they're in dot directories. As
such, `#results` can be 0 even though the scanner found stuff.

Avoid this by also returning candidate count, and not falling back
unless that is 0 as well.