fix: fix failure to resize match listing on VimResized
I noticed we're only getting one callback (ie. for the prompt but not
for the match listing). Fix that by not doing a buffer-local autocmd.

Next problem: the `WinClosed` event doesn't fire reliably, or if it
does, `nvim_del_autocmd()` doesn't always work. Not sure which, but it
happens routinely. So, we hack around that with a group that we clear
before bringing up the UI. At worst, a lingering autocmd is going to
hang around until the next time the user uses Command-T.