feat: add ability for finders to provide an `options` transformer
So that specific finders can override options dynamically/temporarily.
For example, the finders in this commit have been updated to always show
dotfiles, because that is the logical behavior for items that are not on
the filesystem (leading dots only have that special meaning in the
filesystem).

Example: try searching for `netrwbook` (to find `.netrwbook`) in
`:CommandTHelp` before and after this commit.