lint: avoid some "Need check nil" diagnostics
Applied some refactors based on looking through the output of

    :lua vim.diagnostic.setqflist()

For self recursive functions, I had:

    local e = nil

    e = function()
      -- code that calls e()
    end

which was telling me I needed to check nil before calling `e()`.

This doesn't work:

    local e = function()
      -- code that calls e()
    end

(It reports "Undefined global `e`").

This form, however, is the idiomatic one, and suffers from neither
problem:

    local function e()
      -- code that calls e()
    end

It's not a panacea though. For example, it won't help with mutually
recursive functions:

    local function a()
      b() -- Undefined global `b`
    end

    local function b()
      a()
    end

So that's why I'm not doing this everywhere yet. For the latter, like in
`validate.lua`, I still have a bunch of forward declarations to work
around this. Will need a different pattern to avoid those.