feat: add `match_listing` / `truncate` setting
For: https://github.com/wincent/command-t/discussions/429

Some disgusting arithmetic in here, so hopefully it's correct.

Note that this setting takes string values:

- 'beginning'
- 'middle' (the default)
- 'end' (the old behavior prior to this commit)

But also booleans:

- true (same as 'middle')
- false (same as 'end')

And for foolproof-ness, stringy booleans:

- 'true'
- 'false'

Note that 'middle' is a reasonable default because that's what the old
Ruby implementation did (and it wasn't even configurable):

- https://github.com/wincent/command-t/blob/3e937f91260acca0061d884d0ff3c9a521c1fde9/ruby/command-t/lib/command-t/match_window.rb#L478-L487

Also note that when 'end', we don't show an actual ellipsis; we just cut
off the text at the right edge.