chore: stop clang-format spam
Recently started spewing:

    Configuration file(s) do(es) not support C:

errors (one for each file) because I was passing it `*.c` filenames and
it complains that I haven't told it what to do with those.

Tried changing `Language: Cpp` to `Language: C` and it instead
complained:

    Configuration file(s) do(es) not support C++:

ie. it evidently stopped complaining about the `*.c` files, and started
complaining about the `*.h` files, because it considers the latter to be
C++...

You can't do:

    Language: C, Cpp

because it blows up with:

    Error reading .clang-format: Invalid argument
    .clang-format:3:11: error: unknown enumerated scalar

But it seems that dropping the `Language:` entry entirely suffices to
make things work.