fix: handle Unicode properly when computing string lengths
When drawing icons (which are Unicode codepoints with multi-byte
representations), we were running afoul of the fact that all of these
Lua string operations operate on bytes instead of codepoints:

- `#` to determine length
- `string.len`
- `string.format` with the `%-99s` format marker
- `string.sub`

This would cause us to not make full use of the available space: we
would truncate strings too aggressively when inserting a middle
ellipsis, and we wouldn't draw the selected line highlight all the way
to the end of the line.

The fix is to use `vim.str_utfindex` to determine width by counting
codepoints, and `vim.str_byteindex` to figure out the byte index of a
particular codepoint (so that we can truncate the string with
`string.sub`, passing it that byte index).