fix: handle multi-cell glyphs better
Still not perfect, because our truncation calculations don't take glyph
size into consideration:

    local half = math.floor((width - 2) / 2)
    local left = sub(line, 1, half - 2 + width % 2)
    local right = sub(line, 2 - half)
    line = left .. '...' .. right

ie.

- compute half of available space (after subtracting 2 for borders or
  the selected indicator; I don't really remember which).
- grab prefix to show on left
- grab suffix to show on the right
- join with an ellipsis (...)

The "grab" prefix steps don't know about whether the string segments
they are grabbing contain multi-cell glyphs or not. If they do, the
segment will be too long, and when we draw the line in the window, we'll
end up having to truncate at the end as well.

There's a similar problem with aiming to truncate at the beginning.

Fixing this would probably be expensive, so I think I'm going to
tolerate this for now.