refactor: attempt harder to handle edge cases
This has things working correctly for single and multi-cell strings when
`match_listing.truncate` is set to `"beginning"`.

It works correctly for multi-cell strings when `match_listing.truncate`
is set to `"middle"`, but it seems off for single-cell strings (you can
see it flipping back and forth between two lengths as you change the
window width).