feat: implement `traverse` support for `:CommandT`
Rather hacky, so want to test this a bit to make sure it works, probably
refactor it to make it less hacky, and then roll out coverage to the
other file-oriented finder types.

In the end, treating explicit path arguments as overriding the inferred
root directory; ie. `:CommandT foo` or `:CommandT ..` won't do
traversal, but `:CommandT` will.

In all cases, this behaves as if you `cd`'d into the root (either
explicitly specified or found via traversal) ran the command, and the
`cd`'d back. The match listing shows candidates relative to that root.