feat: implement `traverse` support for `:CommandTGit`
Takes a different approach than the one take for `:CommandT`.

In that one, we did a `cd` before scanning, then a `cd` afterwards to
put us back where we started.

In this one, we don't need to `cd` but can instead pass the path to
`git -C`.

In both cases, we need to take the selected item's path, which will be
relative to that other directory, and convert it into a path that Neovim
can open relative to the starting directory.

This, like the other change, is fiddly, complicated, and brittle,
suggesting a need to refactor all of this once the dust has settled.