fix: don't `xrealloc()` so aggressively in `watchman_append_char()`
I'm not sure if I'm forgetting my original reasons for growing the
capacity so much, so there is a risk that this is actually a regression,
but anyway...

Before — if we don't have enough room to append a char, we were using
`+=` to grow the buffer by the current length of the buffer
(`w->length`) and additionally adding `WATCHMAN_DEFAULT_STORAGE` (4096)
bytes for good measure.

After — we just grow the buffer by `WATCHMAN_DEFAULT_STORAGE` (4096)
bytes.

The old pattern would grow the buffer fewer times in the face of many
repeated calls, but it would be more likely to over-allocate.

Example: let's imagine we start with length 4096 and capacity 4096 and
try to append a char — the old code would have grown the buffer to 4096
+ 4096 + 4096, while the new code will grow it to 4096 + 4096.