fix: be more carefully growing the buffer in `watchman_append()`
Another one similar to the parent commit, in the sense that we now
`xrealloc()` less aggressively.

The old code would grow capacity by `w->length` +
`WATCHMAN_DEFAULT_STORAGE` (4096) bytes.

The new code will grow it by `WATCHMAN_DEFAULT_STORAGE` (4096) bytes if
the requested additional capacity is less than 4096 bytes, and
`WATCHMAN_DEFAULT_STORAGE` + the requested additional capacity if it is
more than 4096 bytes.

Note that in the unlikely scenario that `length` was very large, the old
code risks overrunning the buffer. I don't think this was happening in
practice, but you never know.

Example — if `length` was huge (say, 128K), and `w->length` and
`w->capacity` were small (say, 8192), then we would bump `w->capacity`
to 8192 + 8192 + 4096, which is not enough to hold 128K.