fix: avoid potential source of segfaults in watchman scanner
Instead of just leaking and returning `NULL`, the solution is to not
leak (because that's the right thing to do, not necessarily to fix the
crashes) and to implement proper error handling at the call sites. It's
this latter change that might actually fix some segfaults, because under
the load of the benchmarks, it seems like the watchman server isn't
always giving us a valid response.

This is not a definitive fix, however. In my testing, with the this
change, I was able to get the `bin/benchmarks/scanner.lua` script to run
to completion (ie. running all scanners over the test data 20 times
rehearsal + 20 times actual). But after rebuilding, this was no longer
the case; it was back to its old, crashy self. So, evidently there is
still more work to be done here.