fix: actually fix Makefile
Claude was wrong. The parent commit avoids the unnecessary rebuilds, but
it overcorrects and makes it so that we never rebuilt when `CCFLAGS`
changed.

Went back to Stackoverflow, and found this pertinent answer:

> I don't understand how the other solutions are supposed to work. If
> the ARGS target is .PHONY or depends on a .PHONY, then it will always
> be run, right?
>
> Here is my solution using the $(file) function in newer versions of
> gmake:
>
> ...

- https://stackoverflow.com/a/74378629/2103996

But that solution doesn't work either:

```
.PHONY: FORCE

define DEPENDABLE_VAR
$(1):
        echo -n $($(1)) > $(1)
ifneq ("$(file <$(1))","$($(1))")
$(1): FORCE
endif

endef

$(eval $(call DEPENDABLE_VAR,ARGS))

foo: foo.c ARGS
        touch foo
```

It is going to write to `ARGS` every time, updating its timestamp, and
causing `foo` to be built every time.

The final solution implemented in this commit works as follows:

1. When the Makefile is parsed (before anything actually runs), the
   `DEPEND_ON` function is called (with `$(call)`).
2. In the call to `DEPEND_ON`, `$1` is `CCFLAGS`.
3. We define a target with no dependencies, `.make/CCFLAGS`.
4. If and only if the contents of `.make/CCFLAGS` on disk is different
   from the contents of `$CCFLAGS`, we write an updated version back to
   the file.
5. If and only if the contents of `.make/CCFLAGS` on disk is different
   from the contents of `$CCFLAGS`, we add a dependency on the `.FORCE`
   phony target.
6. Via `$(eval)`, the targets we just set up in steps 3, 4, and 5 get
   incorporated into the Makefile.

The end result of this is that, if new CCFLAGS were in effect at parse
time, we'll have these rules, which `make` will then use to decide what
actually gets built:

```
.PHONY: .FORCE
.make/CCFLAGS:
  @if [[ `cat .make/CCFLAGS 2>&1` != "$CCFLAGS" ]]; then \
    printf '%s' "$CCFLAGS" > .make/CCFLAGS; \
  fi
.make/CCFLAGS: .FORCE
```

Make will merge those two targets so the effective result is:

```
.PHONY: .FORCE
.make/CCFLAGS: .FORCE
  @if [[ `cat .make/CCFLAGS 2>&1` != "$CCFLAGS" ]]; then \
    printf '%s' "$CCFLAGS" > .make/CCFLAGS; \
  fi
```

ie. the `.make/CCFLAGS` target _will_ get rebuilt because it depends on
the phony `.FORCE` target. The `if` will succeed and update the file on
disk. Because `commandt.$(DLLEXT)` depends on `.make/CCFLAGS`, that will
get rebuilt too.

On the next run, assuming CCFLAGS are the same, the resulting rules
after parsing and the `$(eval)` call will be:

```
.PHONY: .FORCE
.make/CCFLAGS:
  @if [[ `cat .make/CCFLAGS 2>&1` != "$CCFLAGS" ]]; then \
    printf '%s' "$CCFLAGS" > .make/CCFLAGS; \
  fi
```

Because `.make/CCFLAGS` exists on disk and has no dependencies, it won't
get rebuilt. Because `commandt.$(DELLEXT)` depends on `.make/CCFLAGS`
but the timestamp on the latter is older than the one on the former,
that target will not get rebuilt.